+++ cmd/internal/gengo/go.go
@@ -113,12 +113,7 @@ func runCmd(cmd *base.Command, args []string) {
 		}
 		return nil
 	})
-	modload.LoadModFile()
-	modload.SyncGoMod()
-	if modload.ClassModFile != nil && modload.ClassModFile.Classfile != nil {
-		cl.RegisterClassFileType(modload.ClassModFile.Classfile.ProjExt,
-			modload.ClassModFile.Classfile.WorkExt, modload.ClassModFile.Classfile.PkgPaths...)
-	}
+	modload.Load()
 	runner.GenGo(dir, true, &cl.Config{CacheLoadPkgs: !*flagSlow})
 	errs := runner.Errors()
 	if errs != nil {
