+++ cmd/internal/mod/download.go
@@ -13,7 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package modcmd
+// gop mod download
+package mod
 
 import (
 	"github.com/goplus/gop/cmd/internal/base"
