+++ cmd/internal/mod/mod.go
@@ -13,39 +13,19 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-// gop mod init
-
-package modcmd
+package mod
 
 import (
-	"log"
-
 	"github.com/goplus/gop/cmd/internal/base"
-	"github.com/goplus/gop/cmd/internal/modload"
 )
 
-var cmdInit = &base.Command{
-	UsageLine: "go mod init [module]",
-	Short:     "initialize new module in current directory",
-}
+var Cmd = &base.Command{
+	UsageLine: "gop mod",
+	Short:     "module maintenance",
 
-func init() {
-	cmdInit.Run = runInit
-}
-
-func runInit(cmd *base.Command, args []string) {
-	if len(args) > 1 {
-		log.Fatalf("gop mod init: too many arguments")
-	}
-	var modPath string
-	if len(args) == 1 {
-		modPath = args[0]
-	}
-
-	// modfetch.InitArgs(".", args...)
-	modload.CreateModFile(modPath) // does all the hard work
-	modload.LoadModFile()
-	modload.SyncGoMod()
-	modload.SyncGopMod()
+	Commands: []*base.Command{
+		cmdInit,
+		cmdDownload,
+		cmdTidy,
+	},
 }
