+++ cmd/internal/mod/tidy.go
@@ -13,7 +13,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package modcmd
+// gop mod tidy
+package mod
 
 import (
 	"github.com/goplus/gop/cmd/internal/base"
