+++ cmd/internal/run/run.go
@@ -143,12 +143,7 @@ func runCmd(cmd *base.Command, args []string) {
 	if isDir {
 		srcDir = src
 		gofile = src + "/gop_autogen.go"
-		modload.LoadModFile()
-		modload.SyncGoMod()
-		if modload.ClassModFile != nil && modload.ClassModFile.Classfile != nil {
-			cl.RegisterClassFileType(modload.ClassModFile.Classfile.ProjExt,
-				modload.ClassModFile.Classfile.WorkExt, modload.ClassModFile.Classfile.PkgPaths...)
-		}
+		modload.Load()
 		isDirty = true // TODO: check if code changed
 		if isDirty {
 			pkgs, err = parser.ParseDir(fset, src, nil, parserMode)
@@ -157,12 +152,7 @@ func runCmd(cmd *base.Command, args []string) {
 		}
 	} else {
 		srcDir, file = filepath.Split(src)
-		modload.LoadModFile()
-		modload.SyncGoMod()
-		if modload.ClassModFile != nil && modload.ClassModFile.Classfile != nil {
-			cl.RegisterClassFileType(modload.ClassModFile.Classfile.ProjExt,
-				modload.ClassModFile.Classfile.WorkExt, modload.ClassModFile.Classfile.PkgPaths...)
-		}
+		modload.Load()
 		isGo := filepath.Ext(file) == ".go"
 		if isGo {
 			hash := sha1.Sum([]byte(src))
